[fix] only show approve modal if user has erc20 tokens
