+++ src/components/ApproveTokens/ApproveTokens.js
@@ -1,36 +1,13 @@
 import React from "react";
 import "./ApproveTokens.css";
 import { Trans } from "@lingui/macro";
-import { getTokens } from "config/tokens";
 import { useWeb3React } from "@web3-react/core";
-import useSWR from "swr";
-import { getContract } from "config/contracts";
-import { contractFetcher } from "lib/contracts";
-
-import Reader from "abis/ReaderV2.json";
-import { useInfoTokens } from "domain/tokens";
 import ApproveTokenInput from "components/ApproveTokenInput/ApproveTokenInput";
 import Button from "components/Button/Button";
 
 export default function ApproveTokens(props) {
-  const { chainId, pendingTxns, setPendingTxns, closeApprovalsModal } = props;
-  const { active, account, library } = useWeb3React();
-  const readerAddress = getContract(chainId, "Reader");
-  const tokens = getTokens(chainId);
-  const tokenAddresses = tokens.map((token) => token.address);
-
-  const { data: tokenBalances } = useSWR(active && [active, chainId, readerAddress, "getTokenBalances", account], {
-    fetcher: contractFetcher(library, Reader, [tokenAddresses]),
-  });
-  const { infoTokens } = useInfoTokens(library, chainId, active, tokenBalances);
-
-  let nonZeroBalanceTokens = [];
-  for (let key in infoTokens) {
-    let tokenInfo = infoTokens[key];
-    if (tokenInfo.balance && tokenInfo.balance.gt(0)) {
-      nonZeroBalanceTokens.push(tokenInfo);
-    }
-  }
+  const { chainId, pendingTxns, setPendingTxns, nonZeroBalanceTokens, closeApprovalsModal } = props;
+  const { library } = useWeb3React();
 
   return (
     <div className="Approve-tokens-modal-body">
